renzocastro91/model.py
class Producto:
    def __init__(self, nombre, precio):
        self.nombre = nombre
        self.precio = precio

class Carrito:
    def __init__(self):
        self.productos = []

    def agregar_producto(self, producto):
        self.productos.append(producto)

    def eliminar_producto(self, nombre):
        for producto in self.productos:
            if producto.nombre == nombre:
                self.productos.remove(producto)
                return producto
        return None
    
    def obtener_productos(self):
        return self.productos

    def obtener_total(self):
        total = sum(producto.precio for producto in self.productos)
        return total